requirements.txt
fastapi
uvicorn[standard]
supabase
python-dotenv
python-multipart
huggingface_hub